backend/requirements.txt
asgiref==3.5.2
attrs==22.1.0
autopep8==1.7.0
backports.zoneinfo==0.2.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1
django-allauth==0.50.0
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
drf-spectacular==0.23.1
idna==3.3
importlib-resources==5.9.0
inflection==0.5.1
jsonschema==4.14.0
oauthlib==3.2.0
pkgutil_resolve_name==1.3.10
praw==7.6.0
prawcore==2.3.0
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.4.0
pyrsistent==0.18.1
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tenacity==8.0.1
toml==0.10.2
update-checker==0.18.0
uritemplate==4.1.1
urllib3==1.26.12
websocket-client==1.4.0
zipp==3.8.1
